requirements.txt
mem0ai
qdrant-client
langchain
langchain_community
langchain_core
langchain_google_genai
langchain_groq
google-genai
google-generativeai
fastapi
uvicorn
python-multipart
langgraph
discord.py
logging
python-pptx
python-docx
pytesseract
Pillow
PyMuPDF
cohere
streamlit
